Move CoreProperties to core
